modify tests for Hotel.sol to test for onlyOwner
The tests for non-owner execution were using the `callHotel` function in `WTIndex.sol`, so they were actually testing that `msg.sender` exists in the `hotelsByManager` mapping. This commit changes the tests in `hotel.js` to execute functions directly on the `Hotel.sol` contract with a non-owner, in order to test that it's public functions are protected by the `onlyOwner` modifier.

This commit also adds tests to `WTIndex.js` that test the `callHotel` function directly.